Refactor: Reorganize codebase into logical folder structure
Restructured the DopeGrid package to improve code organization and discoverability:

**DopeGrid Package Structure:**
- `Common/`: Shared utilities (FlipAxis, RotationDegree, SpanBitArray, Pool)
- `Grid/`: Core grid system (IGridShape, GridBoard, IndexedGridBoard, ImmutableGridShape)
- `Inventory/`: Game inventory abstractions (IGameInventory, DefaultGameInventory, SharedData)
- `Editor/`: Editor-only types (EditorGridShape)

**DopeInventory Package Changes:**
- Moved game inventory model to DopeGrid (platform-agnostic)
- Kept UI-specific code in DopeInventory-uGUI (IUIInventory, InventoryView)
- Extracted UI-specific shared data (ISharedUIInventoryData) to UI layer
- Removed Unity dependency from SharedInventoryData

**Benefits:**
- Clearer separation between platform-agnostic grid logic and Unity-specific UI
- Better code discoverability with logical grouping
- Enables potential use of inventory model in non-Unity contexts
- Follows Single Responsibility Principle for package boundaries
