+++ Packages/com.fullmetalbagel.dope-grid/Runtime/Common/Pool.cs
