+++ Packages/com.fullmetalbagel.dope-grid/Runtime/Common/ReadOnlySpanBitArray.cs
