+++ Packages/com.fullmetalbagel.dope-grid/Runtime/Common/RotationDegree.cs.meta
