+++ Packages/com.fullmetalbagel.dope-grid/Runtime/Common/SpanBitArray.cs
