+++ Packages/com.fullmetalbagel.dope-grid/Runtime/Common/SpanBitArrayUtility.cs
