+++ Packages/com.fullmetalbagel.dope-grid/Runtime/Grid/GridShape.cs.meta
