+++ Packages/com.fullmetalbagel.dope-grid/Runtime/Grid/IGridShape.cs.meta
