+++ Packages/com.fullmetalbagel.dope-grid/Runtime/Grid/IReadOnlyGridShape.cs
