+++ Packages/com.fullmetalbagel.dope-grid/Runtime/Grid/ImmutableGridShape.cs
