+++ Packages/com.fullmetalbagel.dope-grid/Runtime/Grid/UnsafeGridShape.cs
