+++ Packages/com.fullmetalbagel.dope-grid/Runtime/Grid/ValueGridShape.cs.meta
