+++ Packages/com.fullmetalbagel.dope-grid/Runtime/Inventory/DefaultGameInventory.cs.meta
