+++ Packages/com.fullmetalbagel.dope-grid/Runtime/Inventory/IGameInventory.cs
