+++ Packages/com.fullmetalbagel.dope-grid/Runtime/Inventory/SharedData.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using JetBrains.Annotations;
-using UnityEngine;
 
 namespace DopeGrid.Inventory;
 
@@ -19,25 +18,6 @@ public interface ISharedInventoryData
     void SetRotation(Guid itemInstanceId, RotationDegree rotation);
 }
 
-public interface ISharedUIInventoryData
-{
-    IList<DraggingItem> DraggingItems { get; }
-
-    [Pure, MustUseReturnValue]
-    Sprite GetSprite(Guid itemInstanceId);
-}
-
-public static class SharedExtension
-{
-    [Pure, MustUseReturnValue]
-    public static ImmutableGridShape GetRotatedShape(this ISharedInventoryData shared, Guid itemInstanceId)
-    {
-        var shape = shared.GetShape(itemInstanceId);
-        var rotation = shared.GetRotation(itemInstanceId);
-        return shape.GetRotatedShape(rotation);
-    }
-}
-
 public class SharedInventoryData : ISharedInventoryData
 {
     public Dictionary<Guid, IGameInventory?> ItemInventoryMap { get; } = new();
@@ -49,11 +29,4 @@ public class SharedInventoryData : ISharedInventoryData
     public ImmutableGridShape GetShape(Guid itemInstanceId) => ItemShapeMap[itemInstanceId];
     public RotationDegree GetRotation(Guid itemInstanceId) => ItemRotationMap[itemInstanceId];
     public void SetRotation(Guid itemInstanceId, RotationDegree rotation) => ItemRotationMap[itemInstanceId] = rotation;
-}
-
-public class SharedUIInventoryData : ISharedUIInventoryData
-{
-    public Dictionary<Guid/*definition id*/, Sprite> ItemSpriteMap { get; } = new();
-    public IList<DraggingItem> DraggingItems { get; } = new List<DraggingItem>();
-    public Sprite GetSprite(Guid itemInstanceId) => ItemSpriteMap[itemInstanceId];
 }
