+++ Packages/com.fullmetalbagel.dope-grid/Runtime/Inventory/SharedData.cs.meta
