+++ Packages/com.fullmetalbagel.dope-inventory-ugui/Runtime/Model/ISharedUIInventoryData.cs
@@ -1,0 +1,21 @@
+using System;
+using System.Collections.Generic;
+using JetBrains.Annotations;
+using UnityEngine;
+
+namespace DopeGrid.Inventory;
+
+public interface ISharedUIInventoryData
+{
+    IList<DraggingItem> DraggingItems { get; }
+
+    [Pure, MustUseReturnValue]
+    Sprite GetSprite(Guid itemInstanceId);
+}
+
+public class SharedUIInventoryData : ISharedUIInventoryData
+{
+    public Dictionary<Guid/*definition id*/, Sprite> ItemSpriteMap { get; } = new();
+    public IList<DraggingItem> DraggingItems { get; } = new List<DraggingItem>();
+    public Sprite GetSprite(Guid itemInstanceId) => ItemSpriteMap[itemInstanceId];
+}
