+++ Packages/com.fullmetalbagel.dope-inventory-ugui/Runtime/Model/ISharedUIInventoryData.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: f1123c74b7894908b7a8d7646849070c
+timeCreated: 1759817556
