+++ Packages/com.fullmetalbagel.dope-inventory-ugui/Runtime/UI/InventoryView.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using UnityEngine;
+﻿using UnityEngine;
 using UnityEngine.EventSystems;
 
 namespace DopeGrid.Inventory
